% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbum.R
\name{getAlbum}
\alias{getAlbum}
\title{GetAlbum}
\usage{
getAlbum(id_album, apikey)
}
\arguments{
\item{id_album}{ID of an artist's album}

\item{apikey}{You API Key provided by MusixMatch}
}
\description{
This function searches for data from an album based on album_id and api key
}
\examples{
apikey="YOUR_API_KEY"
getAlbum(4250417,apikey)
}
