% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLyrics.R
\name{getLyrics}
\alias{getLyrics}
\title{getLyrics}
\usage{
getLyrics(id_track, apikey)
}
\arguments{
\item{apikey}{Your API Key provided by MusixMatch}

\item{idtrack}{Track's ID}
}
\description{
This function gets the lyrics of a track.
}
\examples{
apikey="YOUR_API_KEY")
getlyrics(15953433,apikey)
}
