% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelated.R
\name{getRelated}
\alias{getRelated}
\title{GetRelated}
\usage{
getRelated(artist_id, apikey)
}
\arguments{
\item{artist_id}{ID of the desired artist.}

\item{apikey}{Your API Key provided by MusixMatch.}
}
\description{
This function gets a list of artists somehow related to a given one.
}
\examples{
apikey="YOUR_API_KEY"
getRelated(56,apikey)
}
