% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchArtist.R
\name{searchArtist}
\alias{searchArtist}
\title{searchArtist}
\usage{
searchArtist(artist, apikey)
}
\arguments{
\item{artist}{Artist's name}

\item{apikey}{Your API Key provided by MusixMatch}
}
\description{
This function gets the artist data from MusixMatch database.
}
\examples{
apikey="YOUR_API_KEY"
searchArtist("Taylor Swift",apikey)
}
