% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTrack.R
\name{searchTrack}
\alias{searchTrack}
\title{searchtrack}
\usage{
searchTrack(artist, song, apikey)
}
\arguments{
\item{artist}{Artist's name}

\item{song}{Song's name}

\item{apikey}{Your API Key provided by MusixMatch}
}
\description{
This function searches for track in MusixMatch database.
}
\examples{
apikey="YOUR_API_KEY"
searchTrack("Taylor Swift","Blank Space",apikey)
}
